fix snippets avg rating + switch ratings to snippet store
