+++ client/js/actions/rating-actions.js
@@ -6,21 +6,15 @@ class RatingActions {
   constructor() {
   }
 
-  getAll () {
-    axios.get('/ratings')
-      .then(res => this.dispatch({ok: true, ratings: res}))
-      .catch(err => this.dispatch({ok: false, error: err}));
-  }
-
-  create(rating) {
+  createRating(rating) {
     axios.post('/ratings', rating)
       .then(res =>  this.dispatch({ok: true, rating: res.data}))
       .catch(err => this.dispatch({ok: false, error: err}));
   }
 
   getSnippetRatings (snippet_id) {
-    axios.get('/snippets/' + snippet_id + '/ratings')
-      .then(res => this.dispatch({ok: true, ratings: res.data, snippetId: snippet_id}))
+    axios.get('/snippets/' + snippet_id + '/avg')
+      .then(res => this.dispatch({ok: true, avg: res.data}))
       .catch(err => this.dispatch({ok: false, error: err}));
   }
