+++ client/js/components/ratings/rating-form.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import AltContainer from 'alt/AltContainer.js';
-import RatingStore from '../../stores/rating-store';
+import SnippetStore from '../../stores/snippet-store.js';
 import Rating from './rating';
 
 
@@ -13,8 +13,8 @@ export default class RatingForm extends React.Component {
   render() {
     return (
       <form className="rating-form" style={{margin: '10px', right: 0, float: 'right', position: 'absolute'}}>
-        <AltContainer store={RatingStore}>
-          <Rating snippetId={this.props.snippetId} enabled={this.props.enabled}/>
+        <AltContainer store={SnippetStore}>
+          <Rating snippetId={this.props.snippetId} snippet={this.props.snippet} enabled={this.props.enabled}/>
         </AltContainer>
       </form>
     );
